# apispec_1.json
{
...
  "info": {
    "description": "powered by Flasgger",
    "termsOfService": "/tos",
    "title": "A swagger API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/moderateur/add": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "adresse": {
                  "type": "string"
                },
                "image": {
                  "type": "string"
                },
                "nom": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "prenom": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                },
                "telephone": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Mod\u00e9rateur cr\u00e9\u00e9 avec succ\u00e8s"
          },
          "400": {
            "description": "Donn\u00e9es incompl\u00e8tes pour cr\u00e9er un mod\u00e9rateur"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Ajoute un nouveau mod\u00e9rateur."
      }
    },
    "/api/moderateur/all": {
      "get": {
        "responses": {
          "200": {
            "description": "Liste des mod\u00e9rateurs"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "R\u00e9cup\u00e8re tous les mod\u00e9rateurs enregistr\u00e9s."
      }
    },
    "/api/moderateur/delete/{moderateur_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "moderateur_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Suppression r\u00e9ussie"
          },
          "404": {
            "description": "Mod\u00e9rateur non trouv\u00e9"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Supprime un mod\u00e9rateur de la base de donn\u00e9es."
      }
    },
    "/api/moderateur/login": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "password": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connexion r\u00e9ussie"
          },
          "400": {
            "description": "Donn\u00e9es incompl\u00e8tes"
          },
          "401": {
            "description": "Identifiants invalides"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Connecte un mod\u00e9rateur."
      }
    },
    "/api/moderateur/logout": {
      "post": {
        "responses": {
          "200": {
            "description": "D\u00e9connexion r\u00e9ussie"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "D\u00e9connecte un mod\u00e9rateur."
      }
    },
    "/api/moderateur/protected": {
      "get": {
        "responses": {
          "200": {
            "description": "D\u00e9tails du mod\u00e9rateur connect\u00e9"
          },
          "404": {
            "description": "Mod\u00e9rateur non trouv\u00e9"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "R\u00e9cup\u00e8re les informations du mod\u00e9rateur connect\u00e9."
      }
    },
    "/api/moderateur/register": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "adresse": {
                  "type": "string"
                },
                "image": {
                  "type": "string"
                },
                "nom": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "prenom": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                },
                "telephone": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Mod\u00e9rateur enregistr\u00e9 avec succ\u00e8s"
          },
          "400": {
            "description": "Donn\u00e9es incompl\u00e8tes ou type d'utilisateur invalide"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Inscrit un nouveau mod\u00e9rateur."
      }
    },
    "/api/moderateur/update/{moderateur_id}": {
      "put": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "adresse": {
                  "type": "string"
                },
                "image": {
                  "type": "string"
                },
                "nom": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "prenom": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                },
                "telephone": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mod\u00e9rateur mis \u00e0 jour avec succ\u00e8s"
          },
          "400": {
            "description": "Donn\u00e9es manquantes"
          },
          "404": {
            "description": "Mod\u00e9rateur non trouv\u00e9"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Met \u00e0 jour les informations d'un mod\u00e9rateur."
      }
    },
    "/api/moderateur/{moderateur_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "moderateur_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "D\u00e9tails du mod\u00e9rateur"
          },
          "404": {
            "description": "Mod\u00e9rateur non trouv\u00e9"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "R\u00e9cup\u00e8re un mod\u00e9rateur sp\u00e9cifique via son ID."
      }
    },
...
    "/api/user/add": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "adresse": {
                  "type": "string"
                },
                "nom": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                },
                "telephone": {
                  "type": "string"
                },
                "user_type": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Utilisateur cr\u00e9\u00e9 avec succ\u00e8s"
          },
          "400": {
            "description": "Donn\u00e9es incompl\u00e8tes pour cr\u00e9er un utilisateur"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Ajoute un nouvel utilisateur."
      }
    },
    "/api/user/all": {
      "get": {
        "responses": {
          "200": {
            "description": "Liste des utilisateurs"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "R\u00e9cup\u00e8re tous les utilisateurs enregistr\u00e9s."
      }
    },
    "/api/user/delete/{user_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Suppression r\u00e9ussie"
          },
          "404": {
            "description": "Utilisateur non trouv\u00e9"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Supprime un utilisateur de la base de donn\u00e9es."
      }
    },
    "/api/user/login": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "password": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connexion r\u00e9ussie"
          },
          "400": {
            "description": "Identifiants requis"
          },
          "401": {
            "description": "Authentification \u00e9chou\u00e9e"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Connecte un utilisateur."
      }
    },
    "/api/user/logout": {
      "post": {
        "responses": {
          "200": {
            "description": "D\u00e9connexion r\u00e9ussie"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "D\u00e9connecte un utilisateur."
      }
    },
    "/api/user/protected": {
      "get": {
        "responses": {
          "200": {
            "description": "D\u00e9tails de l'utilisateur connect\u00e9"
          },
          "404": {
            "description": "Utilisateur non trouv\u00e9"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "R\u00e9cup\u00e8re les informations de l'utilisateur connect\u00e9."
      }
    },
    "/api/user/register": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "adresse": {
                  "type": "string"
                },
                "nom": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                },
                "telephone": {
                  "type": "string"
                },
                "user_type": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Enregistrement r\u00e9ussi"
          },
          "400": {
            "description": "Donn\u00e9es incompl\u00e8tes ou type d'utilisateur invalide"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Inscrit un nouvel utilisateur."
      }
    },
    "/api/user/update/{user_id}": {
      "put": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "adresse": {
                  "type": "string"
                },
                "nom": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                },
                "telephone": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Utilisateur mis \u00e0 jour avec succ\u00e8s"
          },
          "400": {
            "description": "Donn\u00e9es manquantes"
          },
          "404": {
            "description": "Utilisateur non trouv\u00e9"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "Met \u00e0 jour les informations d'un utilisateur."
      }
    },
    "/api/user/{user_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "D\u00e9tails de l'utilisateur"
          },
          "404": {
            "description": "Utilisateur non trouv\u00e9"
          },
          "500": {
            "description": "Erreur interne"
          }
        },
        "summary": "R\u00e9cup\u00e8re un utilisateur sp\u00e9cifique via son ID."
      }
    },
...
  },
  "swagger": "2.0"
}